requirements.txt
flask
requests
python-dotenv
python-telegram-bot==20.7
